+++ ShardScript/include/shard/runtime/ConsoleHelper.h
@@ -11,7 +11,7 @@ namespace shard
 	public:
 		static void Write(ObjectInstance* instance);
 		static void Write(bool data);
-		static void Write(int data);
+		static void Write(int64_t data);
 		static void Write(double data);
 		static void Write(wchar_t data);
 		static void Write(const wchar_t* data);
@@ -20,7 +20,7 @@ namespace shard
 		static void WriteLine(ObjectInstance* instance);
 		static void WriteLine();
 		static void WriteLine(bool data);
-		static void WriteLine(int data);
+		static void WriteLine(int64_t data);
 		static void WriteLine(double data);
 		static void WriteLine(wchar_t data);
 		static void WriteLine(const wchar_t* data);
